brahmin_to_tamil_app/requirements.txt.txt
Flask==2.3.2
flask-cors==4.0.0
tensorflow==2.15.0
Pillow==10.3.0
gunicorn==21.2.0

	
